change gdk_threads_enter import from unsafe to safe
Safe imports have two uses: when a function may call back into Haskell, it must
be marked safe; and when a function may block (or take a long time to compute),
it may be marked safe to avoid blocking all other Haskell threads from making
progress. The cost of marking a function safe is decreased (serial)
performance.

Since the typical use case for gdk_threads_enter is to grab a lock in
concurrent code, it is important to ensure that other Haskell threads can make
progress -- especially since one of them may be holding the lock.